Made it so day view is reloading events from main thread always. Resolved a number of warnings of deprecated calls (some are still remaining).

+++ CalendarLib/MGCDayPlannerView.m
@@ -1942,7 +1942,7 @@ static const CGFloat kMaxHourSlotHeight = 150.;
         
         NSUInteger hiddenCount = [self.allDayEventsViewLayout numberOfHiddenEventsInSection:indexPath.section];
         UILabel *label = [[UILabel alloc]initWithFrame:view.bounds];
-        label.text = [NSString stringWithFormat:NSLocalizedString(@"%d more...", nil), hiddenCount];
+        label.text = [NSString stringWithFormat:NSLocalizedString(@"%lu more...", nil), (unsigned long)hiddenCount];
         label.textColor = [UIColor blackColor];
         label.font = [UIFont systemFontOfSize:11];
         label.autoresizingMask = UIViewAutoresizingFlexibleHeight|UIViewAutoresizingFlexibleWidth;
